Solve day9

+++ src/day2.rs
@@ -1,10 +1,10 @@
 use crate::computer::{parse_memory, run};
 
-pub fn first(content: &String) -> i32 {
+pub fn first(content: &String) -> i128 {
     trial(&parse_memory(content), 12, 2)
 }
 
-pub fn second(content: &String) -> i32 {
+pub fn second(content: &String) -> i128 {
     let target = 19690720;
     let initial = parse_memory(content);
     let min_noun = (0..100)
@@ -12,16 +12,16 @@ pub fn second(content: &String) -> i32 {
         .find(|i| trial(&initial, *i, 99) <= target)
         .unwrap();
     let max_noun = (0..100).find(|i| trial(&initial, *i, 0) >= target).unwrap();
-    let verbs = (0..100).collect::<Vec<i32>>();
+    let verbs = (0..100).collect::<Vec<i128>>();
     for noun in min_noun..(max_noun + 1) {
         if let Ok(verb) = verbs.binary_search_by_key(&target, |&v| trial(&initial, noun, v)) {
-            return noun * 100 + verb as i32;
+            return noun * 100 + verb as i128;
         }
     }
     0
 }
 
-fn trial(mem: &Vec<i32>, noun: i32, verb: i32) -> i32 {
+fn trial(mem: &Vec<i128>, noun: i128, verb: i128) -> i128 {
     let mut memory = mem.clone();
     memory[1] = noun;
     memory[2] = verb;
